LSD-60 - /u prefix included in all profile routes -> login required
