+++ src/main/resources/templates/profile.html
@@ -21,7 +21,7 @@
     <div id="content" class="clearfix">
         <div id="userphoto"><img src="images/avatar.png" alt="default avatar"/>
             <h1 id="username"> Your Name </h1>
-            <a type="submit" id="btn-edit" href="/edit-profile" class="btn btn-info btn-lg">
+            <a type="submit" id="btn-edit" href="/u/edit-profile" class="btn btn-info btn-lg">
                 <span class="glyphicon glyphicon-pencil"></span>    Edit Profile</a>
         </div>
         <h1>Your Profile</h1>
