+++ src/main/resources/templates/profile_form.html
@@ -12,7 +12,7 @@
     <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.2.17/angular.min.js"></script>
     <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.2.18/angular-resource.js"></script>
     <script src="https://code.angularjs.org/1.3.0-rc.2/angular-messages.js"></script>
-    <script type="text/javascript" src="js/controller/profileController.js"></script>
+    <script type="text/javascript" src="/js/controller/profileController.js"></script>
 </head>
 <body ng-controller="profileController">
 <div class="container">
@@ -88,8 +88,8 @@
                 </ul>
             </li>
             <li>
-                <a href="/profile" type="button" id="cancel" class="btn btn-info btn-lg">Cancel</a>
-                <a href="/profile" type="button" id="submit" class="btn btn-info btn-lg"
+                <a href="/u/profile" type="button" id="cancel" class="btn btn-info btn-lg">Cancel</a>
+                <a href="/u/profile" type="button" id="submit" class="btn btn-info btn-lg"
                    ng-disabled="profileForm.$invalid" ng-click="saveProfile()">Submit</a>
             </li>
         </ul>
